implement facilitator view
